% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ani2list.R
\name{Ani2list}
\alias{Ani2list}
\title{Convert the AniSpace object into a list}
\usage{
Ani2list(AniObj, verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
A list that contains the information from the AniSpace object.
}
\description{
\code{Ani2list} export information from an AniSpace object to a list object.
}
\examples{
df.list=Ani2list(AniObj=df)

names(df.list)

head(df.list$Pos_info)


}
\keyword{animal}
\keyword{area}
\keyword{export}
\keyword{information}
\keyword{position}
