% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ani2tbl.R
\name{Ani2tbl}
\alias{Ani2tbl}
\title{Convert the AniSpace object into a data frame}
\usage{
Ani2tbl(AniObj, verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
A data frame that contains only the position information from the AniSpace object.
}
\description{
\code{Ani2list} export the position information from an AniSpace object to a data frame object.
}
\examples{
df.tbl=Ani2tbl(AniObj=df)

head(df.tbl)

}
\keyword{export}
\keyword{information}
\keyword{position}
