% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniObj.R
\docType{data}
\name{AniObj}
\alias{AniObj}
\title{Example AniSpace object}
\format{
An S4 object of class \code{AniSpace}.
}
\usage{
AniObj
}
\description{
Example \code{AniSpace} object containing spatio-temporal position information
from dairy cattle monitored using a real-time location system in a
commercial dairy barn.

The object contains individual information, animal positions and polygon
information describing the barn areas. A one-hour period is included to
provide a compact dataset for examples and demonstrations of AniSpace
functions.
}
\keyword{datasets}
