% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaUse.R
\name{areaUse}
\alias{areaUse}
\title{Estimate area utilisation}
\usage{
areaUse(
  AniObj,
  Area = NULL,
  method = "TimeBudget",
  t.lim = 600,
  t.err = t.lim * 0.05,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{Area}{Either \code{NULL}, a character vector containing area identifiers, or a numeric vector containing the indices of the areas to include (\emph{Default: NULL; all areas}).}

\item{method}{A variable indicating the  selected analysis method (\code{InOut} or \code{TimeBudget}) (Default: \emph{method="TimeBudget"}). \code{InOut} reports the entrance and exist time of each individual in each area included in the object. \code{TimeBudget} reports the individuals' time spent in each area.}

\item{t.lim}{A numeric variable indicating the time limit for considering an area occupied (\emph{Default: 600}).}

\item{t.err}{A numeric variable indicating the time error than some positions could fall outside the area within the \code{t.lim} interval (\emph{Default: 5\% of t.lim}).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}
}
\value{
Descriptive statistics of the individual area usage periods and time budgets.
}
\description{
\code{areaUse} estimates individuals' area occupation periods and time budgets statistics.
}
\examples{
# Estimate individual movement information
df.TB=areaUse(AniObj,method="TimeBudget")
head(df.TB)

df.IO=areaUse(AniObj,method="InOut")
head(df.IO)

}
