% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{as.data.frame,AniSpace-method}
\alias{as.data.frame,AniSpace-method}
\title{as.data.frame method for AniSpace}
\usage{
\S4method{as.data.frame}{AniSpace}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{An \code{AniSpace} object.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A data frame containing the individual IDs and position information.
}
\description{
Converts the position information stored in an \code{AniSpace} object into a data frame.
}
