% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{as.list,AniSpace-method}
\alias{as.list,AniSpace-method}
\title{as.list method for AniSpace}
\usage{
\S4method{as.list}{AniSpace}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{An \code{AniSpace} object.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
A list containing the information stored in the \code{AniSpace} object.
}
\description{
Converts an \code{AniSpace} object into a list containing its slots.
}
