% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_matrix.R
\name{cross3}
\alias{cross3}
\title{Cross product of two 3-vectors}
\usage{
cross3(a, b)
}
\arguments{
\item{a, b}{Numeric vectors of length 3.}
}
\value{
A numeric vector of length 3.
}
\description{
Cross product of two 3-vectors
}
\keyword{internal}
