% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edist.R
\name{edist}
\alias{edist}
\title{Estimate euclidean distance}
\usage{
edist(x1 = NULL, y1 = NULL, x2 = NULL, y2 = NULL)
}
\arguments{
\item{x1}{set number one of temporal-position information in the X axis}

\item{y1}{set number one of temporal-position information in the Y axis}

\item{x2}{set number two of temporal-position information in the X axis}

\item{y2}{set number two of temporal-position information in the Y axis}
}
\value{
a vector with the Euclidean distance information corresponding to the input values
}
\description{
\code{edist} estimate euclidean distance between temporal-position information
}
\examples{
set.seed(123)

x1=runif(10,0,10)
y1=runif(10,0,10)
x2=runif(10,0,10)
y2=runif(10,0,10)

euclidean.dist=edist(x1,y1,x2,y2)
head(euclidean.dist)

}
\keyword{internal}
