% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_roles.R
\name{ensure_members}
\alias{ensure_members}
\title{Are these members of the level?}
\usage{
ensure_members(data, members, level, arg, call = rlang::caller_env())
}
\arguments{
\item{data}{An aniframe.}

\item{members}{Values the caller named.}

\item{level}{The identity variable they should belong to.}

\item{arg}{Name of the argument they came from.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Are these members of the level?
}
\keyword{internal}
