% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAniSpace.R
\name{filterAniSpace}
\alias{filterAniSpace}
\title{Filtering temporal-spatial information}
\usage{
filterAniSpace(
  AniObj,
  NIDs = NULL,
  IDs = NULL,
  TimeWindow = NULL,
  Area = NULL,
  soft.boundaries = 0,
  bring.back = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{NIDs}{Numeric variable or vector containing the assigned index of the individuals (\emph{Default: NULL}).}

\item{IDs}{Character variable or vector containing the identification of the individuals (\emph{Default: NULL}).}

\item{TimeWindow}{A numeric vector indicating the limits (minimum and maximum) of the time window to be maintained (\emph{Default: NULL}).}

\item{Area}{A character variable or vector indicating the areas for which time/position data will be preserved (\emph{Default: ALL}).}

\item{soft.boundaries}{A non-negative numeric value specifying the maximum distance outside an area's boundary at which positions are still retained (\emph{Default: 0}).}

\item{bring.back}{A logical value indicating whether positions located outside the selected area should be moved inside the boundaries of that single area (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
A filtered animal position object
}
\description{
\code{filterAniSpace} filtering the individual temporal-spatial information.
}
\examples{
# Filter five individuals
df.ID.filt=filterAniSpace(AniObj,NIDs=c(1:5))
df.ID.filt

# Filter a specific time window
Tmin=as.numeric(as.POSIXct("2020-10-16 11:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"))
Tmax=as.numeric(as.POSIXct("2020-10-16 11:59:59", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"))
df.Time.filt=filterAniSpace(AniObj,TimeWindow=c(Tmin,Tmax))
df.Time.filt

# Filter positions from the resting area
df.Area.filt=filterAniSpace(AniObj,Area=names(AniObj@Area)[c(1:5)])
df.Area.filt


}
