% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.AniSpace.R
\name{interpolate.AniSpace}
\alias{interpolate.AniSpace}
\title{Interpolate temporal-spatial information}
\usage{
interpolate.AniSpace(
  AniObj,
  method = "makima",
  TRes = 1,
  TCap = 120,
  tails = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{method}{A character variable indicating the selected interpolation method (Default: \emph{method="makima"}): \code{prev} stads for interpolate using the previous position known, \code{linear} stands for a linear interpolation method, \code{spline} stands for a interpolation method based on drawing the individual trajectory based on a third degree polynom, \code{makima} stands for a interpolation method based on drawing the individual trajectory based on the modified akima algoritm and, \code{mixed} stands for a mixed interpolation method /under development/.}

\item{TRes}{A numeric variable in seconds indicating the output frequency of the position data (Default: \emph{TRes=1}).}

\item{TCap}{A numeric variable indicating time threshold in seconds about of which interpolation will not be performed (Default: \emph{TCap=120}).}

\item{tails}{A logical variable indicating whether the start and end of the day are filled with the first and last known positions (TRUE) or not (FALSE) (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
A interpolated animal position object
}
\description{
\code{interpolate.AniSpace} interpolating individual temporal-spatial information.
}
\examples{
df.inter=interpolate.AniSpace(df, method="makima", TRes=1)
df.inter


}
\keyword{animal}
\keyword{information}
\keyword{interpolation}
\keyword{position}
\keyword{spatial}
\keyword{temporal}
