% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.Area.R
\name{load.Area}
\alias{load.Area}
\title{Load areas' information}
\usage{
load.Area(AniObj, AreaObj, verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{AreaObj}{A list containing additional information for the areas.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
An AniSpace object
}
\description{
\code{load.Area()} load the information on the distribution of the barn areas.
}
\examples{
data(cows)
AniObj=load.Space(cows$positions)
AniObj=load.Info(AniObj,InfObj=cows$animals)

# Convert rectangular area information to polygons
AreaObj=square2poly(cows$areas)

# Add area information to the AniSpace object
AniObj=load.Area(AniObj,AreaObj)

}
\keyword{area}
\keyword{information}
\keyword{load}
