% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.Info.R
\name{load.Info}
\alias{load.Info}
\title{Load individuals' information}
\usage{
load.Info(AniObj, InfObj, ID.col = "ID", verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{InfObj}{A data frame containing additional information for the individuals (one row per ID).}

\item{ID.col}{A character variable specifying the name of the column in \code{InfObj} that contains the animal IDs (\emph{Default: "ID"}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
An AniSpace object
}
\description{
\code{load.Info()} adds per-individual information (metadata) to an existing AniSpace object
by matching records in \code{InfObj} to \code{AniObj} using the identifier column \code{ID.col}.
}
\examples{
data(cows)
AniObj=load.Space(cows$positions)

# Add individual information to the AniSpace object
AniObj=load.Info(AniObj,InfObj=cows$animals)
AniObj

}
