% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{merge,AniSpace,AniSpace-method}
\alias{merge,AniSpace,AniSpace-method}
\title{Merge two AniSpace objects}
\usage{
\S4method{merge}{AniSpace,AniSpace}(x, y, verbose = TRUE, ...)
}
\arguments{
\item{x}{An \code{AniSpace} object.}

\item{y}{An \code{AniSpace} object to be merged with \code{x}.}

\item{verbose}{A logical value indicating whether informative messages are
printed (\emph{Default: TRUE}).}

\item{...}{Additional arguments. Currently ignored.}
}
\value{
An \code{AniSpace} object containing the merged information.
}
\description{
Merges two \code{AniSpace} objects, combining their individual information,
positions and area information. When the same individual is present in both
objects, position information is combined and duplicated time points are
removed, retaining the information from \code{x}.

Spatial-interaction, utilisation-distribution and utilisation-distribution
similarity results are removed after merging because they must be
re-estimated from the merged position information.
}
\examples{
# Filter the AniSpace Object
df1=filterAniSpace(AniObj,NIDs=c(1:5))
df2=filterAniSpace(AniObj,NIDs=c(1:5))

# Merge the AniSpace Object
df.merge=merge(df1,df2)
df.merge

}
