% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.AniSpace.R
\name{movement.AniSpace}
\alias{movement.AniSpace}
\title{Analyse individual movement information}
\usage{
movement.AniSpace(
  AniObj,
  k.means = FALSE,
  k.mov = NULL,
  k.max = 10,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{k.means}{A logical variable indicating whether the Kmeans cluster analysis will be executed (TRUE) or not (FALSE) (\emph{Default: FALSE}). \code{k.means} will be automatically set to TRUE if \code{k.mov} value is provided. The Kmeans cluster analysis uses the speed and turning angle information from all individuals in the \code{AniObj} to assign \code{k.mov} behaviours to the different areas detailed in the object.}

\item{k.mov}{A numeric variable indicating the number of behaviours used to estimate similarity between areas applying a Kmeans cluster analysis (\emph{Default: NULL}).}

\item{k.max}{A numeric variable indicating the maximum number of clusters tested on the optimisation of the K-means cluster analysis (\emph{Default: 10}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
Descriptive information about how individuals move and behave in different areas.
}
\description{
\code{movement.AniSpace} estimates the distance walked, speed and the turning angles per individual, and performs a cluster analysis of the areas included in the \code{AniObj} object to classify different behaviours.
}
\examples{
# Estimate individual movement information
df.mov=movement.AniSpace(df)
head(df.mov)

# Group different areas based on movement information fixing number of behaviour to 2 (k.mov=2)
df.mov2=movement.AniSpace(df,k.mov=2)
head(df.mov2[[2]])

}
\keyword{individual}
\keyword{information}
\keyword{movement}
