% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{names,AniSpace-method}
\alias{names,AniSpace-method}
\title{Names method for AniSpace}
\usage{
\S4method{names}{AniSpace}(x)
}
\arguments{
\item{x}{An \code{AniSpace} object.}
}
\value{
A character vector containing the slot names of the object.
}
\description{
Returns the slot names of an \code{AniSpace} object.
}
