% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{plot,AniSpace,missing-method}
\alias{plot,AniSpace,missing-method}
\title{Plot method for AniSpace}
\usage{
\S4method{plot}{AniSpace,missing}(x, y, ...)
}
\arguments{
\item{x}{An \code{AniSpace} object.}

\item{y}{Missing.}

\item{...}{Additional arguments passed to \code{plotAniSpace()}.}
}
\description{
Plots the spatial position information contained in an \code{AniSpace} object.
}
