% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSI.R
\name{plotSI}
\alias{plotSI}
\title{Plot pairwise spatial-interaction matrices and networks}
\usage{
plotSI(
  AniObj,
  NIDs = NULL,
  IDs = NULL,
  type = "matrix",
  method = "time",
  int.thr = NULL,
  group = NULL,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An object of class \code{AniSpace} containing a pairwise
spatial-interaction adjacency matrix stored in the \code{SI} slot.}

\item{NIDs}{A numeric vector indicating the numerical identifiers of the
individuals to include in the plot. When \code{NULL}, all individuals are
included (\emph{Default: NULL}).}

\item{IDs}{A character vector indicating the identifiers of the individuals
to include in the plot. When \code{NULL}, selection is based on \code{NIDs}
(\emph{Default: NULL}).}

\item{type}{A character value specifying the type of plot. Available options
are \code{"matrix"} and \code{"network"}. \code{"matrix"} plots the spatial-interaction
adjacency matrix using \code{image}, whereas \code{"network"} plots the corresponding
interaction network using \code{igraph} (\emph{Default: "matrix"}).}

\item{method}{A variable indicating the type of proximity interactions to export (\code{time}, \code{nint} or \code{mean}) (\emph{Default: "time"}).
\code{time} will export the duration per pair of animals in closer proximity than the specified \code{dist.thr} threshold.
\code{nint} will export the number of instances where each pair of animals got within the specified \code{dist.thr} threshold of each other.
\code{mean} estimates the average of time interacted per instance.}

\item{int.thr}{A positive numeric value specifying the numeric threshold that
be used to dichotomise the interaction information (\emph{Default: NULL}).}

\item{group}{A vector containing the group assigned to each selected
individual. Groups are represented by different vertex colours in the
network plot. When \code{NULL}, all individuals belong to the same group
(\emph{Default: NULL}).}

\item{verbose}{A logical variable indicating whether the function will print
relevant information (\code{TRUE}) or not (\code{FALSE}) (\emph{Default: TRUE}).}
}
\value{
Plot the adjacency matrix when \code{type = "matrix"} or the
corresponding \code{igraph} object when \code{type = "network"}.
}
\description{
\code{plotSI} plots pairwise spatial-interaction information stored in
the \code{SI} slot of an \code{AniSpace} object.
}
\examples{
# Estimate spatial interactions
df.SI=spatialInt(AniObj)
df.SI@SI$M[1:5,1:5]

# Plot the spatial-interaction adjacency matrix
plotSI(df.SI,type="matrix")

# Plot the spatial-interaction network
plotSI(df.SI,NIDs=c(3:7),type="network")

}
