% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_utils.R
\name{rad_to_deg}
\alias{rad_to_deg}
\title{Convert radians to degrees}
\usage{
rad_to_deg(x)
}
\arguments{
\item{x}{Numeric vector of angles (radians).}
}
\value{
Numeric vector of angles expressed in degrees.
}
\description{
Convert radians to degrees
}
