% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Space.R
\name{read.Space}
\alias{read.Space}
\title{Read individual temporal-spatial information}
\usage{
read.Space(
  path,
  type = "csv",
  header = TRUE,
  col.names = NULL,
  Time.col = "Time",
  ID.col = "ID",
  x.col = "x",
  y.col = "y",
  TRes = 1,
  Temp.sort = TRUE,
  na.strings = "NA",
  verbose = TRUE
)
}
\arguments{
\item{path}{A string containing the location and full name of the file to load.}

\item{type}{A variable indicating the type of file to be loaded: \code{csv} (comma (,) separated), \code{json} or \code{json.gz} (\emph{Default: csv}).}

\item{header}{A logical variable specifying if the animal information file contains headers  (\emph{Default: TRUE}).}

\item{col.names}{A character string containing the new names of the columns in the file; they are assigned automatically after the file has been read.}

\item{Time.col}{A character variable specifying the name of the column containing the temporal information in epoch format in the file (\emph{Default: "Time"}).}

\item{ID.col}{A character variable specifying the name of the column containing the animal ID information in the file (\emph{Default: "ID"}).}

\item{x.col}{A character variable specifying the name of the column containing the spatial x-axis information in the file (\emph{Default: "x"}).}

\item{y.col}{A character variable specifying the name of the column containing the spatial y-axis information in the file (\emph{Default: "y"}).}

\item{TRes}{A numeric variable indicating the time resolution/frequency of the positions (\emph{Default: 1 sec}).}

\item{Temp.sort}{A variable indicating whether the \code{load.Space} function should (TRUE) or should not (FALSE) sort the spatio-temporal information by \code{ID.col} and \code{Time.col} (\emph{Default: TRUE}).}

\item{na.strings}{A string to interpret as missing values, passed to read.csv() (\emph{Default: "NA"}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
An AniSpace object
}
\description{
\code{read.Space()} load the individual temporal-spatial information into the R environment.
}
\examples{
data(cows)

# Create a temporary position file
file=tempfile(fileext=".csv")
utils::write.csv(cows$positions,file,row.names=FALSE)

# Read position information into an AniSpace object
AniObj=read.Space(file,type="csv")
AniObj

# Remove temporary file
unlink(file)

}
