% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords_rotate.R
\name{rotation_for}
\alias{rotation_for}
\title{The rotation for one subject at one moment}
\usage{
rotation_for(primary, secondary, target)
}
\arguments{
\item{primary}{The vector between the first two alignment points.}

\item{secondary}{The vector to the third, or \code{NULL}.}

\item{target}{Where they should end up.}
}
\value{
A 3x3 rotation matrix.
}
\description{
The rotation for one subject at one moment
}
\keyword{internal}
