% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{show,AniSpace-method}
\alias{show,AniSpace-method}
\title{Show method for AniSpace}
\usage{
\S4method{show}{AniSpace}(object)
}
\arguments{
\item{object}{An \code{AniSpace} object.}
}
\description{
Displays a summary of the information contained in an \code{AniSpace} object.
}
