% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinter.R
\name{sinter}
\alias{sinter}
\title{Interpolate individual data}
\usage{
sinter(l, int.obj, method)
}
\arguments{
\item{l}{A list object containing sorted numeric variables indicating the positions that need to be interpolated.}

\item{int.obj}{A data frame object containing sorted temporal-spatial information from an unique individual.}

\item{method}{A character variable indicating the interpolation method.}
}
\value{
A list of interpolated positions
}
\description{
\code{sinter} Interpolate individual temporal-spatial information. This function is implemented in the \code{interpolate.AniPos} function.
}
\examples{
AniPos.id.inter=sinter(l,int.obj,method="linear")
head(AniPos.id.inter)

}
\keyword{cluster}
\keyword{distance}
\keyword{individual}
\keyword{information}
\keyword{spatial}
\keyword{temporal}
