% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smooth individual movement trajectories}
\usage{
smooth(AniObj, smooth.wind = 3, verbose = TRUE)
}
\arguments{
\item{AniObj}{An object of class \code{AniSpace} containing the individuals' spatio-temporal position information.}

\item{smooth.wind}{A positive odd integer specifying the size of the running-median window. (Default: \emph{smooth.wind=3}).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}
}
\value{
The modified \code{AniSpace} object with smoothed individual trajectories.
}
\description{
\code{smooth} smooths individual movement trajectories by applying a running median to the \code{x} and \code{y} coordinates stored in an \code{AniSpace} object.
}
\examples{
# Smooth individual trajectories using a 3-position running median
df.smooth=smooth(AniObj,smooth.wind=3,verbose=FALSE)
df.smooth

# Compare original and smoothed coordinates for the first individual
head(data.frame(
  x=AniObj@Pos[[1]]$x,
  x.smooth=df.smooth@Pos[[1]]$x,
  y=AniObj@Pos[[1]]$y,
  y.smooth=df.smooth@Pos[[1]]$y
))

}
