% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialint.AniSpace.R
\name{spatialint.AniSpace}
\alias{spatialint.AniSpace}
\title{Estimate spatial interactions among individuals}
\usage{
spatialint.AniSpace(
  AniObj,
  method = "time",
  dist.thr = 250,
  time.thr = 600,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{method}{A variable indicating the type of proximity interactions to export (\code{time}, \code{nint} or \code{all}) (\emph{Default: "time"}). \code{time} will export the duration spent by each pair of animals that are closer to each other than the specified \code{dist.thr} threshold. \code{nint} will record the number of instances where each pair of animals got within the specified \code{dist.thr} threshold of each other. \code{all} estimates both methods.}

\item{dist.thr}{A numeric variable indicating the distance threshold to be considered when estimating spatial interactions (\emph{Default: 250 cm}).}

\item{time.thr}{A numeric variable indicating the time gap between spatial interactions to be considered a new interaction (\emph{Default: 600 s}).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}
}
\value{
An adjacency matrix containing information about pairwise spatial interactions.
}
\description{
\code{spatialint.AniSpace} estimates the pairwise spatial interactions between the individuals.
}
\examples{
# Estimate spatial interactions
df.AdjMat=spatialint.AniSpace(df)
AniPos.AdjMat[1:5,1:5]

# Spatial interactions accounted as number of interactions
AniPos.AdjMat=spatialint.AniSpace(df, method="all")
AniPos.AdjMat[1:5,1:5]

}
\keyword{animal}
\keyword{information}
\keyword{interactions}
\keyword{spatial}
