% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialInt.R
\name{spatialInt}
\alias{spatialInt}
\title{Estimate pairwise spatial interactions among individuals}
\usage{
spatialInt(
  AniObj,
  method = "all",
  dist.thr = 250,
  time.thr = 600,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An object of class \code{AniSpace} containing the individuals' spatio-temporal position information.}

\item{method}{A variable indicating the type of proximity interactions to export (\code{time}, \code{nint} or \code{all}) (\emph{Default: "all"}).
\code{time} will export the duration per pair of animals in closer proximity than the specified \code{dist.thr} threshold.
\code{nint} will export the number of instances where each pair of animals got within the specified \code{dist.thr} threshold of each other.
\code{all} estimates both methods.}

\item{dist.thr}{A positive numeric value specifying the maximum distance between two individuals for them to be considered interacting. It must use the same spatial units as the coordinates stored in \code{AniObj.} (\emph{Default: 250} (cm)).}

\item{time.thr}{A positive numeric value specifying the time interval used to distinguish separate proximity interactions. It must use the same time units as the Time values stored in \code{AniObj} (\emph{Default: 600} (s)).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}
}
\value{
The modified \code{AniSpace} object, with the estimated pairwise spatial-interaction adjacency matrix (SI) stored in \code{AniObj}.
}
\description{
\code{spatialInt} estimates pairwise proximity interactions among individuals based on their spatial coordinates and a specified distance threshold.
}
\examples{
# Estimate spatial interactions
df.SI=spatialInt(AniObj)
df.SI@SI$M[1:5,1:5]

}
