% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.spikes.R
\name{spikes.AniSpace}
\alias{spikes.AniSpace}
\title{Filtering and smoothing spikes in the temporal-spatial information}
\usage{
spikes.AniSpace(
  AniObj,
  keep.spikes = FALSE,
  max.speed = NULL,
  perc.speed = 0.99,
  smooth.spikes = TRUE,
  smooth.wind = 41,
  smooth.edges = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{max.speed}{A numeric variable indicating the maximum speed allow (distance/temporal units, e.g. m/s). Example: \emph{max.speed=50} (Default: \emph{max.speed=NULL}).}

\item{perc.speed}{A numeric variable E(0,1) indicating the top percentile of the global speed distribution that should be used as \code{max.speed} to filter the spikes in the position data (Default: \emph{perc.speed=0.99}).}

\item{smooth.wind}{A numeric variable indicating the size of the sliding window applied on \code{smooth.spikes}. The number must me odd (Default: \emph{smooth.wind=41}).}

\item{smooth.edges}{A logical variable indicating whether the first and last \code{smooth.spikes} positions should be reallocated to the first and last smoothed positions (Default: \emph{smooth.edges=TRUE}).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}

\item{method}{A variable indicating the selected approach (\code{filter} or \code{smooth}) (Default: \emph{method="smooth"}). \code{filter} remove position that surpass the speed threshold limit (\code{max.speed}). \code{smooth} smooth the spikes by applying a running median sliding window of size equal to \code{smooth.wind}.}
}
\value{
A spikes filtered and/or smoothed animal position object
}
\description{
\code{spikes.AniSpace} filtering or smoothing spikes from the individual temporal-spatial information using a speed threshold.
}
\examples{
# Filter spikes
df.spikes.filt =spikes.AniSpace(df, method="filter")

# Smooth spikes
df.smooth.filt =spikes.AniSpace(df, method="smooth")

}
\keyword{animal}
\keyword{filter}
\keyword{information}
\keyword{noise}
\keyword{position}
\keyword{smooth}
\keyword{spatial}
\keyword{spikes}
\keyword{temporal}
