% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{spikes}
\alias{spikes}
\title{Filtering and smoothing spikes in the temporal-spatial information}
\usage{
spikes(
  AniObj,
  method = "smooth",
  smooth.wind = 5,
  max.speed = NULL,
  perc.speed = 0.99,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An \code{AniSpace} object containing the spatio-temporal information of the individuals.}

\item{method}{A variable indicating the selected approach (\code{filter} or \code{smooth}) (Default: \emph{method="smooth"}). \code{filter} remove position that surpass the speed threshold limit (\code{max.speed}). \code{smooth} smooth the spikes by applying a running median sliding window of size equal to \code{smooth.wind}.}

\item{smooth.wind}{A numeric variable indicating the size of the sliding window applied on \code{smooth.spikes}. The number must me odd (Default: \emph{smooth.wind=5}).}

\item{max.speed}{A numeric variable indicating the maximum speed allow (distance/temporal units, e.g. m/s). Example: \emph{max.speed=50} (Default: \emph{max.speed=NULL}).}

\item{perc.speed}{A numeric variable E(0,1) indicating the top percentile of the global speed distribution that should be used as \code{max.speed} to filter the spikes in the position data (Default: \emph{perc.speed=0.99}).}

\item{verbose}{A logical variable indicating whether the function will print relevant information (TRUE) or not (FALSE) (\emph{Default: TRUE}).}
}
\value{
A spikes filtered and/or smoothed \code{AniSpace} object
}
\description{
\code{spikes} filtering or smoothing spikes from the individual temporal-spatial information using a speed threshold.
}
\examples{
# Filter positions identified as spikes
df.spikes=spikes(AniObj,method="filter",perc.speed=0.99,verbose=FALSE)
df.spikes

# Compare the number of positions before and after filtering
c(
  original=sum(vapply(AniObj@Pos,function(p) length(p$Time),numeric(1))),
  filtered=sum(vapply(df.spikes@Pos,function(p) length(p$Time),numeric(1)))
)

# Smooth positions identified as spikes
df.smooth=spikes(AniObj,method="smooth",smooth.wind=5,
                 perc.speed=0.99,verbose=FALSE)
df.smooth

}
