% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprog.R
\name{sprog}
\alias{sprog}
\title{Loading bar}
\usage{
sprog(i, n)
}
\arguments{
\item{i}{loop number}

\item{n}{total number of loops}
}
\value{
print a loading bar to show progress
}
\description{
\code{sprog} displays a loading bar to show progress
}
\examples{
pbar=sprog(50,100)
pbar

}
\keyword{bar}
\keyword{loading}
\keyword{progress}
