% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square2poly.R
\name{square2poly}
\alias{square2poly}
\title{Transform areas' information to polygon structure}
\usage{
square2poly(
  AreaObj,
  Area.col = "Area",
  x1.col = "x1",
  x2.col = "x2",
  y1.col = "y1",
  y2.col = "y2"
)
}
\arguments{
\item{AreaObj}{A data frame containing information for the  barn areas in (x1,x2,y1,y2 format).}

\item{Area.col}{A character variable specifying the name of the column containing the Area identification in the file.}

\item{x1.col}{A character variable specifying the name of the column containing the x1 position information in the file.}

\item{x2.col}{A character variable specifying the name of the column containing the x2 position information in the file.}

\item{y1.col}{A character variable specifying the name of the column containing the y1 position information in the file.}

\item{y2.col}{A character variable specifying the name of the column containing the y2 position information in the file.}
}
\value{
list of polygons
}
\description{
\code{square2poly()} tranform the information on the distribution of the barn areas to polygons structure.
}
\examples{
data(cows)

# Convert rectangular area information to polygon format
AreaObj=square2poly(cows$areas)

# Inspect the first polygon
AreaObj[[1]]

}
\keyword{area}
\keyword{information}
\keyword{transform}
