% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.Area.R
\name{stats.Area}
\alias{stats.Area}
\title{Descriptive statistics of the facility drawings}
\usage{
stats.Area(AniObj, graphs = FALSE, verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{graphs}{A logical variable indicates whether the graphs showing the descriptive statistics will be reported back (TRUE) or not (FALSE) (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
Descriptive statistics of the areas detailed in the AniSpace object.
}
\description{
\code{stats.Area} estimates the descriptive statistics of the areas detailed in the AniSpace object.
}
\examples{
# Estimate descriptive statistics for the defined areas
df.area=stats.Area(AniObj,graphs=FALSE,verbose=FALSE)
head(df.area)

# Estimate descriptive statistics and generate the area plot
df.area.graph=stats.Area(AniObj,graphs=TRUE,verbose=FALSE)
head(df.area.graph$DescriptiveStats)

# Display the stored area plot
df.area.graph$graphs[[1]]

}
