% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.Still.R
\name{stats.Still}
\alias{stats.Still}
\title{Estimate accuracy of the positioning system}
\usage{
stats.Still(
  AniObj,
  NIDs = NULL,
  IDs = NULL,
  percentile = 0.99,
  graphs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{NIDs}{Numeric. Vector containing the assigned index of the individuals (\emph{Default: NULL}).}

\item{IDs}{Character. Vector containing the identification of the individuals (\emph{Default: NULL}).}

\item{percentile}{A numeric variable indicating the percentiles to be evaluated (Default: \emph{percentile=0.99}).}

\item{graphs}{A logical variable indicates whether the filtered position will be depicted by individual (TRUE) or not (FALSE) (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
Estimated accuracy of the system
}
\description{
\code{stats.Still} estimate accuracy of the positioning system using stand-still devices
}
\examples{
# Estimate positioning accuracy for selected stand-still devices
df.still=stats.Still(AniObj,NIDs=c(8:17),percentile=0.99,
                     verbose=FALSE)
head(df.still)

# Estimate accuracy for multiple percentiles and generate a plot
df.still.graph=stats.Still(AniObj,NIDs=c(8:17),
                           percentile=c(0.50,0.95,0.99),
                           graphs=TRUE,verbose=FALSE)
head(df.still.graph$stats)

# Display the stored accuracy plot
df.still.graph$graphs

}
