% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stime.R
\name{stime}
\alias{stime}
\title{Estimate time left}
\usage{
stime(i, n, st)
}
\arguments{
\item{i}{loop number}

\item{n}{total number of loops}

\item{st}{time at which the procedure started}
}
\value{
displays the time required to complete a procedure
}
\description{
\code{stime} estimates the time required to complete a procedure
}
\examples{
st=Sys.time()
Sys.sleep(5)
tt=stime(50,100,st)
tt

}
\keyword{estimate}
\keyword{left}
\keyword{time}
