% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords_translate.R
\name{translate_coords_keypoint}
\alias{translate_coords_keypoint}
\title{Translate coordinates relative to a keypoint}
\usage{
translate_coords_keypoint(data, to_keypoint)
}
\arguments{
\item{data}{An aniframe in a Cartesian coordinate system.}

\item{to_keypoint}{A character string naming a keypoint to place at the
origin. All other coordinates become relative to it.}
}
\description{
Translate coordinates relative to a keypoint
}
\keyword{internal}
