% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_coords_translate.R
\name{translate_coords_vector}
\alias{translate_coords_vector}
\title{Translate coordinates relative to coordinates}
\usage{
translate_coords_vector(data, to_x, to_y, to_z = NULL)
}
\arguments{
\item{data}{An aniframe in a Cartesian coordinate system.}

\item{to_x}{A numeric x-coordinate: either one value, or one per time point.}

\item{to_y}{A numeric y-coordinate: either one value, or one per time point.}

\item{to_z}{A numeric z-coordinate, for three-dimensional data: either one
value, or one per time point.}
}
\description{
Translate coordinates relative to coordinates
}
\keyword{internal}
