% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-class.R
\docType{class}
\name{AniSpace-class}
\alias{AniSpace-class}
\title{AniSpace class}
\description{
An S4 class to represent the spatio-temporal data of multiple individuals.
}
\section{Slots}{

\describe{
\item{\code{NIDs}}{Numeric. Vector containing the assigned index of the individuals.}

\item{\code{IDs}}{Character. Vector containing the identification of the individuals.}

\item{\code{Info}}{List. Indexed information about the individuals.}

\item{\code{TLim}}{Numeric. Contains information about the temporal range covered by the positions (Index=1s).}

\item{\code{TRes}}{Numeric. Maximum temporal resolution of the positions (e.g. smallest time step).}

\item{\code{Pos}}{List. Indexed positions (x,y,z) of the individuals.}

\item{\code{Area}}{List. Information about the spatial area available for the individuals (Polygon format).}
}}

\examples{
new("AniSpace",
    NIDs = c(1,2),
    IDs  = c("A","B"),
    Info = list(Trait1 = runif(2), Trait2 = runif(2)),
    TLim = as.numeric(as.POSIXct(c("2021-01-01","2021-01-02"), tz="UTC")),
    TRes = 1,
    Pos  = list(as.matrix(data.frame(x = runif(43200), y = runif(43200))),
                as.matrix(data.frame(x = runif(43200), y = runif(43200)))),
    Area = list(A = as.matrix(data.frame(x = runif(5), y = runif(5))),
                B = as.matrix(data.frame(x = runif(5), y = runif(5))))))

}
