% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anispace-package.R
\docType{package}
\name{anispace}
\alias{anispace-package}
\alias{anispace}
\title{anispace package}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An R package providing spatial transformation methods for movement data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://animovement.dev/anispace/}
  \item \url{https://github.com/animovement/anispace/}
  \item Report bugs at \url{https://github.com/animovement/anispace/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

Authors:
\itemize{
  \item Mikkel Roald-Arbøl \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})
}

}
\keyword{internal}
