% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAniSpace.R
\name{filterAniSpace}
\alias{filterAniSpace}
\title{Filtering temporal-spatial information}
\usage{
filterAniSpace(
  AniObj,
  NIDs = NULL,
  IDs = NULL,
  TimeWindow = NULL,
  Area = NULL,
  soft.boundaries = 0,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{NIDs}{Numeric variable or vector containing the assigned index of the individuals (\emph{Default: NULL}).}

\item{IDs}{Character variable or vector containing the identification of the individuals (\emph{Default: NULL}).}

\item{TimeWindow}{A numeric vector indicating the limits (minimum and maximum) of the time window to be maintained (\emph{Default: NULL}).}

\item{Area}{A character variable or vector indicating the areas for which time/position data will be preserved (\emph{Default: ALL}).}

\item{soft.boundaries}{A numeric variable indicating the distance threshold allowed outside the boundaries that will be brought back to the boundaries. The position reported after this threshold will be filtered (e.g. \emph{soft.boundaries=2}; in the same unit of measurement that corresponds to the data (e.g. metres, centimetres, etc.) (\emph{Default: 0}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
A filtered animal position object
}
\description{
\code{filterAniSpace} filtering the individual temporal-spatial information.
}
\examples{
# Filter five individuals
df.ID.filt=filterAniSpace(df,NIDs=c(1:5))
df.ID.filt

# Filter a specific time window
Tmin=as.numeric(as.POSIXct("2020-10-16 11:00:00", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"))
Tmax=as.numeric(as.POSIXct("2020-10-16 11:59:59", format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC"))
df.Time.filt=filterAniSpace(df,TimeWindow=c(Tmin,Tmax))
df.Time.filt

# Filter positions from the resting area
df.Area.filt=filterAniSpace(df,Area=names(df@Area)[c(1:5)])
df.Area.filt


}
\keyword{animal}
\keyword{filter}
\keyword{information}
\keyword{position}
\keyword{spatial}
\keyword{temporal}
