% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.Area.R
\name{load.Area}
\alias{load.Area}
\title{Load areas' information}
\usage{
load.Area(AniObj, AreaObj, ID.col = "ID", verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{AreaObj}{A list containing additional information for the areas.}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
An AniSpace object
}
\description{
\code{load.Area()} load the information on the distribution of the barn areas.
}
\examples{
# Create AniSpace object
df=read.Space(path="data/positions.csv", type="csv")

# Add manually the information
cf=list(
list(ID="0",
     coords=matrix(c(0,   0, 3340,  0, 3340,2902, 0,2902, 0,0), ncol=2, byrow=TRUE, dimnames=list(NULL,c("x","y"))),
     color="beige"),
list(ID="1",
     coords=matrix(c(0,2902, 630,2902, 630,8738, 0,8738, 0,2902), ncol=2, byrow=TRUE, dimnames=list(NULL,c("x","y"))),
     color="green"
))

# Reading the information from a file
bf=read.csv("data/barnplan.csv", header=T)
cf=square2poly(bf)

# Adding information to the AniSpace object
df=load.Info(AniObj=df, AreaObj=cf)
df

}
\keyword{area}
\keyword{information}
\keyword{load}
