% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.Space.R
\name{load.Space}
\alias{load.Space}
\title{Transform individual spatio-temporal information into an AniSpace object}
\usage{
load.Space(
  PosObj,
  Time.col = "Time",
  ID.col = "ID",
  x.col = "x",
  y.col = "y",
  TRes = 1,
  Temp.sort = TRUE
)
}
\arguments{
\item{PosObj}{A data frame containing the spatio-temporal information of the individuals.}

\item{Time.col}{A character variable specifying the name of the column containing the temporal information in epoch format in the file (\emph{Default: "Time"}).}

\item{ID.col}{A character variable specifying the name of the column containing the animal ID information in the file (\emph{Default: "ID"}).}

\item{x.col}{A character variable specifying the name of the column containing the spatial x-axis information in the file (\emph{Default: "x"}).}

\item{y.col}{A character variable specifying the name of the column containing the spatial y-axis information in the file (\emph{Default: "y"}).}

\item{TRes}{A numeric variable indicating the time resolution/frequency of the positions (\emph{Default: 1 sec}).}

\item{Temp.sort}{A variable indicating whether the \code{load.Space} function should (TRUE) or should not (FALSE) sort the spatio-temporal information by \code{ID.col} and \code{Time.col} (\emph{Default: TRUE}).}
}
\value{
An AniSpace object
}
\description{
\code{load.Space()} transforms a data frame containing individual spatio-temporal information into the AniSpace format.
}
\examples{
df=load.Space(data.frame)
df

}
\keyword{animal}
\keyword{information}
\keyword{spatial}
\keyword{transform}
