% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_cartesian.R
\name{map_to_cartesian}
\alias{map_to_cartesian}
\title{Map from polar to Cartesian coordinates}
\usage{
map_to_cartesian(data)
}
\arguments{
\item{data}{an aniframe with polar coordinates}
}
\value{
an aniframe with Cartesian coordinates
}
\description{
Map from polar to Cartesian coordinates
}
