% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_cylindrical.R
\name{map_to_cylindrical}
\alias{map_to_cylindrical}
\title{Map from Cartesian to cylindrical coordinates}
\usage{
map_to_cylindrical(data)
}
\arguments{
\item{data}{movement data frame with Cartesian coordinates}
}
\value{
movement data frame with cylindrical coordinates
}
\description{
Map from Cartesian to cylindrical coordinates
}
