% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_polar.R
\name{map_to_polar}
\alias{map_to_polar}
\title{Map from Cartesian to polar coordinates}
\usage{
map_to_polar(data)
}
\arguments{
\item{data}{An aniframe in a Cartesian coordinate system.}
}
\value{
An aniframe with \code{rho} and \code{phi} in place of \code{x} and \code{y}.
}
\description{
Map from Cartesian to polar coordinates
}
\examples{
af <- anicore::example_anipoint(
  n_obs = 5, n_individuals = 1, n_keypoints = 1
)
map_to_polar(af)
}
\seealso{
Other coordinate systems:
\code{\link[=map_to_cartesian]{map_to_cartesian()}},
\code{\link[=map_to_cylindrical]{map_to_cylindrical()}},
\code{\link[=map_to_spherical]{map_to_spherical()}}
}
\concept{coordinate systems}
