% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_spherical.R
\name{map_to_spherical}
\alias{map_to_spherical}
\title{Map from Cartesian to spherical coordinates}
\usage{
map_to_spherical(data)
}
\arguments{
\item{data}{A data frame/tibble containing columns \code{x}, \code{y}, \code{z}}
}
\value{
Same data frame with columns \code{rho}, \code{theta}, \code{phi}
}
\description{
Map from Cartesian to spherical coordinates
}
