% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAniSpace.R
\name{plotAniSpace}
\alias{plotAniSpace}
\title{Plot AniSpace position information}
\usage{
plotAniSpace(AniObj, NIDs = NULL, IDs = NULL, Area = NULL, ...)
}
\arguments{
\item{AniObj}{An AniSpace object}

\item{NIDs}{Numeric. Vector containing the assigned index of the individuals (\emph{Default: first}).}

\item{IDs}{Character. Vector containing the identification of the individuals (\emph{Default: first}).}

\item{Area}{Numeric. Information about the areas available for the individuals (Polygon format) (\emph{Default: all}).}
}
\value{
Plot AniSpace position information
}
\description{
\code{plotAniSpace()} plot AniSpace position information
}
\examples{
validate(df)

plot.AniSpace(df, IDs="2428706")

plot.AniSpace(df, IDs="2428706", Area=names(df@Area)[c(1:7)])

}
