% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_utils.R
\name{polar_to_x}
\alias{polar_to_x}
\title{Convert polar radius to Cartesian x‑coordinate}
\usage{
polar_to_x(rho, phi)
}
\arguments{
\item{rho}{numeric vector of radial distances}

\item{phi}{numeric vector of azimuth angles (radians)}
}
\value{
numeric vector of x‑coordinates
}
\description{
Convert polar radius to Cartesian x‑coordinate
}
