% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_utils.R
\name{polar_to_y}
\alias{polar_to_y}
\title{Convert polar radius to Cartesian y‑coordinate}
\usage{
polar_to_y(rho, phi)
}
\arguments{
\item{rho}{numeric vector of radial distances}

\item{phi}{numeric vector of azimuth angles (radians)}
}
\value{
numeric vector of y‑coordinates
}
\description{
Convert polar radius to Cartesian y‑coordinate
}
