% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{rebase}
\alias{rebase}
\title{Rebase AniSpace object}
\usage{
rebase(AniObj)
}
\arguments{
\item{AniObj}{An AniSpace object}
}
\description{
\code{rebase()} Reorganizes or redefines the reference system of an \code{AniSpace} object,
aligning positional or temporal data to a new baseline.
}
\examples{
df=rebase(df)

}
