% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniSpace-methods.R
\name{show,AniSpace-method}
\alias{show,AniSpace-method}
\title{\enumerate{
\item Show method for AniSpace
}}
\usage{
\S4method{show}{AniSpace}(object)
}
\arguments{
\item{AniObj}{An AniSpace object}
}
\description{
\enumerate{
\item Show method for AniSpace
}
}
