% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.Pos.R
\name{stats.Pos}
\alias{stats.Pos}
\title{Descriptive statistics of the temporal-spatial information}
\usage{
stats.Pos(AniObj, graphs = FALSE, verbose = TRUE)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{graphs}{A logical variable indicates whether the graphs showing the descriptive statistics will be reported back (TRUE) or not (FALSE) (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
Descriptive statistics of the temporal-spatial information detailed in the AniSpace object.
}
\description{
\code{stats.Pos} estimates the descriptive statistics of the temporal and spatial information contained within the position object at an individual level.
}
\examples{
df.stats.Pos=stats.Pos(AniObj=df, graphs=TRUE)

head(df.stats.Pos$DescriptiveStats)

df.stats.Pos$graphs[[1]]

}
\keyword{descriptive}
\keyword{information}
\keyword{position}
\keyword{spatial}
\keyword{statistics}
