% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.still.R
\name{stats.still}
\alias{stats.still}
\title{Estimate accuracy of the positioning system}
\usage{
stats.still(
  AniObj,
  NIDs = NULL,
  IDs = NULL,
  percentile = 0.99,
  graphs = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{AniObj}{An AniSpace object containing the spatio-temporal information of the individuals.}

\item{NIDs}{Numeric. Vector containing the assigned index of the individuals (\emph{Default: NULL}).}

\item{IDs}{Character. Vector containing the identification of the individuals (\emph{Default: NULL}).}

\item{percentile}{A numeric variable indicating the percentiles to be evaluated (Default: \emph{percentile=0.99}).}

\item{graphs}{A logical variable indicates whether the filtered position will be depicted by individual (TRUE) or not (FALSE) (\emph{Default: FALSE}).}

\item{verbose}{A logical variable specifying whether to print informative messages (\emph{Default: TRUE}).}
}
\value{
Estimated accuracy of the system
}
\description{
\code{stand.still} estimate accuracy of the positioning system using stand-still devices
}
\examples{
# Estimate accuracy of the positioning system using stand-still devices
StandStill.acc=stand.still(df,keep. NIDs=c(1:10))

}
\keyword{accuracy}
\keyword{devices}
\keyword{estimation}
\keyword{stand-still}
\keyword{system}
