% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_utils.R
\name{unwrap_angle}
\alias{unwrap_angle}
\title{Remove constrain for angles to keep within [0, 2π)}
\usage{
unwrap_angle(x)
}
\arguments{
\item{x}{Numeric vector of angles (radians).}
}
\value{
Numeric vector of the same length.
}
\description{
Unwraps any numeric vector from the interval [0, 2π).
}
