% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate AniSpace object's information}
\usage{
validate(AniObj)
}
\arguments{
\item{AniObj}{Performs a series of integrity and quality-control checks on an
\code{AniSpace} object or dataset, ensuring that positional, animal,
and area data meet expected structural and logical criteria.}
}
\description{
\code{validate()} Verifies and Consistency Checks for AniSpace Objects
}
\examples{
validate(df)

}
