% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate AniSpace object's information}
\usage{
validate(AniObj)
}
\arguments{
\item{AniObj}{An AniSpace object}
}
\description{
\code{validate()} validate AniSpace object's information
}
\examples{
validate(df)

}
